+++ frontend/todolist/src/component/groups/groups.jsx
@@ -1,10 +1,10 @@
-import React from 'react';
+import React, { memo } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faPlus } from '@fortawesome/free-solid-svg-icons';
 import styles from './groups.module.css';
 import Group from '../group/group';
 
-const Groups = ({ groups, handleGroup, onPopupClick }) => {
+const Groups = memo(({ groups, handleGroup, onPopupClick }) => {
   const onAddGroup = () => {
     onPopupClick();
   };
@@ -24,6 +24,6 @@ const Groups = ({ groups, handleGroup, onPopupClick }) => {
       </section>
     </div>
   );
-};
+});
 
 export default Groups;
