+++ frontend/todolist/src/component/header/header.jsx
@@ -1,7 +1,23 @@
-import React from 'react';
+import React, { memo } from 'react';
+import styles from './header.module.css';
+import mainTitle from '../../image/mainTheme.png';
 
-const Header = (props) => {
-  return <h1>header</h1>;
-};
+const Header = memo(({ setIsLogin }) => {
+  const onLogOut = () => {
+    localStorage.clear('token');
+    setIsLogin(false);
+  };
+  return (
+    <header className={styles.header}>
+      <img src={mainTitle} alt='mainTitle' className={styles.mainTitle} />
+      <div>
+        <button className={styles.logoutBtn}>MyName</button>
+        <button onClick={onLogOut} className={styles.logoutBtn}>
+          Log out
+        </button>
+      </div>
+    </header>
+  );
+});
 
 export default Header;
