+++ frontend/todolist/src/component/header/header.module.css
@@ -1,0 +1,43 @@
+.header {
+  position: absolute;
+  height: 60px;
+  width: 100%;
+  top: 0px;
+  left: 0px;
+
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.mainTitle {
+  margin-left: 20px;
+  width: auto;
+  height: 40px;
+  cursor: pointer;
+  transition: all 0.2s ease-in;
+}
+.mainTitle:hover {
+  transform: scale(1.03);
+}
+
+.logoutBtn {
+  height: 40px;
+  border-style: none;
+  margin-right: 10px;
+  padding: 10px 10px;
+  border-radius: 20px;
+
+  -webkit-box-shadow: 10px 10px 21px -15px rgba(0, 0, 0, 0.75);
+  -moz-box-shadow: 10px 10px 21px -15px rgba(0, 0, 0, 0.75);
+  box-shadow: 10px 10px 21px -15px rgba(0, 0, 0, 0.75);
+
+  font-size: 15px;
+
+  cursor: pointer;
+  transition: all 0.1s ease-in;
+}
+
+.logoutBtn:hover {
+  transform: scale(1.02);
+}
